refactor: (#57) Main board/comments 기능 UI 구조 분리
