+++ src/pages/main/model/types.ts
@@ -1,23 +1,1 @@
 export type MainTab = 'ROOM' | 'LUNCH' | 'RANKING' | 'BOARD';
-
-export interface MainBoardPost {
-  id: number;
-  category: 'FREE' | 'REVIEW' | 'INFO' | 'TALK';
-  title: string;
-  author: string;
-  summary: string;
-  content: string;
-  likedCount: number;
-  commentCount: number;
-  createdAt: string;
-}
-
-export interface MainBoardComment {
-  id: number;
-  postId: number;
-  author: string;
-  content: string;
-  likedCount: number;
-  createdAt: string;
-  isMine: boolean;
-}
